+++ ReleaseNotes/release_notes.md
@@ -1,8 +1,9 @@
 ## New features:
   1. add new test/example `test_fitting_simfit7` to compre simultenoud fit versus fit with constraints 
   1. `fit1d` : add suffin to the name for automatically created backgronus component 
-
-
+  1. add `Ostap::Math::hotelling` function to estgimate Hotelling t2-statistics 
+  1. make use of `Ostap::Math::hotelling` function in reweighting tests  
+ 
 ## Backward incompatible: 
 
 ## Bug fixes: